added templating, packaging, docs
now we compile the .yml files with jinja2 before running them
through the yaml parser. This allows us to specify full directory
structures in the .yml file, rather than having to specify a dir
structure and a root directory into which to place that structure.

added package. It appears to work with easy_install but not pip (?!)

improved README and internal help messages.